spec-363: separate memex-search retrieval / ranking / ref-building / formatting [P3]
Behaviour-preserving decomposition of the ~1572 LOC memex-search.ts god-module
(sol-7 of audit spec-345, under umbrella spec-354). The four mixed concerns move
into focused modules under services/memex-search/ — code moved VERBATIM, no
SQL/logic rewrites, so search results stay byte-identical (same RRF ordering,
same canonical refs, same formatted markdown):

- types.ts       public + internal row types
- refs.ts        canonical-path building + docType→kind map
- time.ts        toIso / toMillis timestamp helpers
- retrieval.ts   the six FTS/vector SQL arms, handle lookup, slug +
                 open-comment loads, scope + vector-distance config
- ranking.ts     RRF merge, accumulator, snippet pickers
- formatting.ts  markdown renderer, WHO/WHEN bylines, truncate

memex-search.ts becomes the thin entry/orchestration layer (searchMemex,
resolveJumpTo, resolveAssignedSpecs) and re-exports the full public surface, so
every importer keeps importing the same names from the same path with zero edits.

Tests: search + downstream suites 97 passed; full server suite 4289 passed
(1 pre-existing env-gated spec-129 failure, unrelated). tsc error set identical
to baseline (no new errors).
