+++ packages/server/src/services/memex-search.ts
@@ -48,15 +48,56 @@
 // needed here; the predicate is hard-wired into every query. The canonical
 // predicate is `AND d.is_demo IS NOT TRUE` (column is NOT NULL DEFAULT false,
 // but IS NOT TRUE is robust against any legacy NULL and reads as the intent).
+//
+// ── spec-363 sol-7: this file is now the thin ENTRY/orchestration layer. ──
+// The four concerns it used to mix have moved to focused modules under
+// memex-search/: retrieval (the FTS/vector SQL arms + handle lookup + slug /
+// open-comment loads), ranking (the RRF merge), refs (canonical-path building),
+// and formatting (the markdown renderer). This file wires them together
+// (searchMemex / resolveJumpTo / resolveAssignedSpecs) and re-exports the public
+// surface so every importer keeps importing the same names from the same path.
 
 import { sql } from "drizzle-orm";
-import { timeAgo, capitalizeDisplayName } from "@memex/shared";
 import { db } from "../db/connection.js";
-import {
-  resolveEmbeddingProvider,
-  type EmbeddingProvider,
-} from "./embedding-provider.js";
+import { resolveEmbeddingProvider } from "./embedding-provider.js";
 import { listSpecsAssignedToUser } from "./doc-assignees.js";
+import { buildDocPath, kindForDocType } from "./memex-search/refs.js";
+import { mergeWithRrf } from "./memex-search/ranking.js";
+import {
+  attachOpenComments,
+  inScopeDocTypes,
+  loadMemexSlugs,
+  lookupByHandle,
+  resolveMaxVectorDistance,
+  runDecisionFts,
+  runDecisionVector,
+  runIssueFts,
+  runIssueVector,
+  runSectionFts,
+  runSectionVector,
+} from "./memex-search/retrieval.js";
+import type {
+  DecisionRow,
+  IssueRow,
+  MemexSearchHit,
+  SectionRow,
+} from "./memex-search/types.js";
+
+// Re-export the public surface (types + the formatter) so existing importers —
+// agent/tool-specs.ts, routes/search.ts, mcp/formatters.ts, the test suites —
+// keep importing the same names from "./memex-search.js" with zero edits
+// (spec-363 dec-1).
+export type {
+  FormatOptions,
+  MatchingSection,
+  MemexSearchHit,
+  MemexSearchKind,
+  SearchMemexOptions,
+  SearchStrategy,
+} from "./memex-search/types.js";
+export { formatSearchResults } from "./memex-search/formatting.js";
+
+import type { SearchMemexOptions } from "./memex-search/types.js";
 
 const HANDLE_REGEX = /^(spec|std|doc)-\d+$/i;
 
@@ -80,362 +121,6 @@ const NUMBER_JUMP_SCORE: Record<"spec" | "std" | "doc", number> = {
   std: 0.8,
   doc: 0.7,
 };
-
-// Reciprocal-rank-fusion constant. 60 is the canonical default from the
-// original Cormack/Clarke 2009 paper. Lower k weights top ranks more heavily,
-// higher k flattens the curve. Keep at 60 unless we have measurements that
-// disagree.
-const RRF_K = 60;
-
-// Relevance floor for the semantic (vector) arms (spec-64 i-1). pgvector's
-// `<=>` is cosine distance in [0, 2] — 0 = identical, ~1 = orthogonal /
-// unrelated, 2 = opposite. The vector arms order by distance and LIMIT, but
-// without a ceiling they return their nearest neighbours HOWEVER far away, so a
-// low-signal query (e.g. a person's name with no lexical match) surfaces
-// unrelated sections that don't contain the query terms at all. Any vector hit
-// at or beyond this distance is treated as "not actually related" and dropped.
-//
-// Only the vector arm is floored — FTS hits require a lexeme match (`@@`) so
-// they're inherently relevant, and a doc that also matched FTS still surfaces
-// via that arm even if its vector row is filtered. The RRF merge is unchanged.
-//
-// The right value is embedding-model-specific (this default is tuned for
-// openai-text-embedding-3-large @1536, the prod provider). It is overridable
-// per-env via MEMEX_SEARCH_MAX_VECTOR_DISTANCE (and per-call via
-// SearchMemexOptions.maxVectorDistance) so it can be tuned on INT without a
-// code change before prod.
-const DEFAULT_MAX_VECTOR_DISTANCE = 0.65;
-
-function resolveMaxVectorDistance(explicit?: number): number {
-  if (typeof explicit === "number" && Number.isFinite(explicit)) return explicit;
-  const env = process.env.MEMEX_SEARCH_MAX_VECTOR_DISTANCE;
-  if (env) {
-    const parsed = Number(env);
-    if (Number.isFinite(parsed) && parsed > 0) return parsed;
-  }
-  return DEFAULT_MAX_VECTOR_DISTANCE;
-}
-
-export type SearchStrategy = "handle" | "fts" | "vector";
-
-// Per b-34 D-2: phase-1 scope. `kind` is optional — omit to search every
-// kind. Tasks, comments, execution-plan sections deferred (easy to add as
-// additional kinds later). Issues (spec-112 t-4) join the decision arm as a
-// second non-section kind — same RRF FTS+vector machinery, no new search
-// infra (s-4): they ride their own table's title/body + embedding column.
-export type MemexSearchKind = "spec" | "standard" | "document" | "decision" | "issue";
-
-// Maps the user-facing kind enum onto the underlying docType column values.
-// `'document'` is the broad "free-form" bucket (per CLAUDE.md the `documents`
-// table covers all docTypes; the kind enum is a narrower user-facing
-// vocabulary). 'execution_plan' is omitted from kind=document since Spec
-// users probably don't want execution-plan sections in search results — they
-// can still be added as a separate kind later if demand surfaces.
-const DOC_TYPES_BY_KIND: Record<Exclude<MemexSearchKind, "decision" | "issue">, string[]> = {
-  spec: ["spec"],
-  standard: ["standard"],
-  document: ["document", "adr", "runbook"],
-};
-
-// All section docTypes in scope when `kind` is omitted (decisions are handled
-// by the separate decision arm).
-const ALL_SECTION_DOC_TYPES = [
-  ...DOC_TYPES_BY_KIND.spec,
-  ...DOC_TYPES_BY_KIND.standard,
-  ...DOC_TYPES_BY_KIND.document,
-];
-
-export interface MatchingSection {
-  id: string;
-  sectionType: string;
-  title: string | null;
-  content: string;
-  /** Which search method surfaced this section first. */
-  matchedVia: SearchStrategy;
-  /** WHO — best-effort display name of who last touched THIS section (spec-259
-   *  ac-9). Carried per-section so a multi-section doc hit surfaces each
-   *  matched section's own creator, not just the doc-level latest. Same SQL
-   *  COALESCE(actor_name, users.name, users.email) resolution as the doc-level
-   *  WHO. null when nothing resolved. Wire/structured value — kept absolute. */
-  authorName: string | null;
-  /** WHEN — ISO-8601 last-modified of THIS section (spec-259 ac-9). Absolute
-   *  ISO on the structured object; rendered relative via timeAgo(). */
-  lastUpdatedAt: string | null;
-}
-
-export interface MemexSearchHit {
-  /** Internal UUID — never rendered. The path is the public identifier. */
-  id: string;
-  kind: MemexSearchKind;
-  /** Canonical URL path for this hit (no scheme/host). For decisions:
-   *  `<ns>/<mx>/<docTypePath>/<docHandle>/decisions/dec-N`.
-   *  For docs: `<ns>/<mx>/<docTypePath>/<docHandle>`. */
-  path: string;
-  title: string;
-  /** Status / phase. For docs: docStatus. For decisions: decisionStatus. */
-  status: string;
-  /** Aggregated rank score after RRF merge. Higher = better. */
-  score: number;
-  /** Strategies that contributed at least one hit. */
-  strategies: SearchStrategy[];
-  /** Sections that matched (only populated for kind != 'decision'). */
-  matchingSections: MatchingSection[];
-  /** Snippet for decision hits (matched chunk, ≤ 300 chars). */
-  decisionSnippet?: string;
-  /** Search method that surfaced this decision (only for kind='decision'). */
-  decisionMatchedVia?: SearchStrategy;
-  /** Snippet for issue hits (matched chunk, ≤ 300 chars). spec-112 t-4. */
-  issueSnippet?: string;
-  /** Search method that surfaced this issue (only for kind='issue'). */
-  issueMatchedVia?: SearchStrategy;
-  /** bug | todo — surfaced in the issue hit heading so a search reader can
-   *  tell a bug from a todo without opening it (only for kind='issue'). */
-  issueType?: string;
-  /** Parent doc UUID. For section/doc hits this equals `id`. For decision
-   *  hits this is the parent Spec's UUID. Internal — never rendered, but
-   *  used to detect self-hits when the caller passes `currentDocId` to the
-   *  formatter. */
-  parentDocId: string;
-  /** WHO — best-effort display name of who authored / last touched this hit
-   *  (spec-285 dec-1). For decision/section hits this is the denormalised
-   *  std-32 `actor_name` where present (rename-proof); for document and issue
-   *  hits, the `created_by_user_id` resolved to `users.name ?? users.email`.
-   *  null when nothing resolved (a system write, a deleted user, a legacy
-   *  unattributed row). Navigation-only lanes (jump / assigned) leave it null. */
-  authorName: string | null;
-  /** WHEN — ISO-8601 timestamp of when this hit was last changed (spec-285
-   *  dec-2): the row's last-modified where it has one (`doc_sections`/`issues`
-   *  `updated_at`), falling back to `created_at` where it does not (decisions
-   *  carry no `updated_at`). null only when no timestamp was selected
-   *  (navigation-only lanes). */
-  lastUpdatedAt: string | null;
-  /** Open (unresolved) comments anchored on this hit's doc (spec-259 ac-12). A
-   *  lightweight indicator only — count + oldest open comment's age — never the
-   *  comment content. `oldestCreatedAt` is absolute ISO on the structured
-   *  object; rendered relative via timeAgo(). Omitted (undefined) when the hit
-   *  has zero open comments, so the formatter renders no indicator line. */
-  openComments?: { count: number; oldestCreatedAt: string };
-}
-
-export interface SearchMemexOptions {
-  /** Restrict to one entity kind. Omit for all. */
-  kind?: MemexSearchKind;
-  /** Cap on returned hits. Default 8 — tighter than standards-search to
-   *  protect agent context (per b-34 D-4). */
-  limit?: number;
-  /** Force off vector search even when a provider is configured (e.g. test
-   *  mode). */
-  disableVector?: boolean;
-  /** Inject a deterministic provider; tests use this to avoid API calls. */
-  provider?: EmbeddingProvider | null;
-  /** Include archived / paused content. Default false. */
-  includeArchived?: boolean;
-  /** Max cosine distance for a vector hit to count as relevant — the semantic
-   *  relevance floor (spec-64 i-1). Vector hits at or beyond this distance are
-   *  dropped so a low-signal query doesn't surface unrelated nearest-neighbour
-   *  sections. Defaults to MEMEX_SEARCH_MAX_VECTOR_DISTANCE env, else
-   *  DEFAULT_MAX_VECTOR_DISTANCE. FTS hits are never floored. */
-  maxVectorDistance?: number;
-  /** Exclude hits whose section's doc_id (or decision's parent doc_id)
-   *  matches this UUID. The in-app agent binds the current doc here so
-   *  search results don't include the very Spec being edited (the
-   *  agent already has it in its Document Context system block). MCP
-   *  callers omit this; unset = no filter. */
-  excludeDocId?: string;
-}
-
-interface SectionRow {
-  section_id: string;
-  doc_id: string;
-  section_type: string;
-  section_title: string | null;
-  section_content: string;
-  doc_handle: string;
-  doc_title: string;
-  doc_status: string;
-  doc_type: string;
-  // spec-285: WHO/WHEN. `author_name` is computed in SQL — the section's
-  // denormalised actor_name (std-32) preferred, else the parent doc's resolved
-  // creator. `updated_at` is the section's own last-modified.
-  author_name: string | null;
-  updated_at: string | Date;
-  rank?: number; // FTS ts_rank
-  distance?: number; // vector cosine distance
-}
-
-interface DecisionRow {
-  decision_id: string;
-  doc_id: string;
-  doc_handle: string;
-  doc_title: string;
-  doc_type: string;
-  dec_seq: number;
-  dec_title: string;
-  dec_context: string | null;
-  dec_resolution: string | null;
-  dec_status: string;
-  // spec-285: WHO/WHEN. `author_name` = decision's denormalised actor_name
-  // (std-32) preferred, else the actor_user_id resolved to a user. `created_at`
-  // is the only timestamp decisions carry (no updated_at — dec-2 fallback).
-  author_name: string | null;
-  created_at: string | Date;
-  rank?: number;
-  distance?: number;
-}
-
-interface IssueRow {
-  issue_id: string;
-  doc_id: string;
-  doc_handle: string;
-  doc_title: string;
-  doc_type: string;
-  issue_seq: number;
-  issue_title: string;
-  issue_body: string | null;
-  issue_type: string;
-  issue_status: string;
-  // spec-285: WHO/WHEN. Issues carry no actor_name (std-32 hasn't reached the
-  // issues table) — `author_name` is the created_by_user_id resolved to a user.
-  // `updated_at` is the issue's own last-modified.
-  author_name: string | null;
-  updated_at: string | Date;
-  rank?: number;
-  distance?: number;
-}
-
-// Resolved Memex slug parts — needed to build canonical paths.
-interface MemexSlugs {
-  namespace_slug: string;
-  memex_slug: string;
-}
-
-async function loadMemexSlugs(memexId: string): Promise<MemexSlugs | null> {
-  const rows = (await db.execute(sql`
-    SELECT n.slug AS namespace_slug, m.slug AS memex_slug
-    FROM memexes m
-    INNER JOIN namespaces n ON n.id = m.namespace_id
-    WHERE m.id = ${memexId}
-    LIMIT 1
-  `)) as unknown as MemexSlugs[];
-  return rows[0] ?? null;
-}
-
-// docType → URL path segment. Matches the routing convention: specs at
-// /specs, standards at /standards, free-form at /docs, execution plans at
-// /execution-plans.
-function docTypeToPathSegment(docType: string): string {
-  if (docType === "spec") return "specs";
-  if (docType === "standard") return "standards";
-  if (docType === "execution_plan") return "execution-plans";
-  return "docs"; // document, adr, runbook, etc.
-}
-
-function buildDocPath(slugs: MemexSlugs, docType: string, handle: string): string {
-  return `${slugs.namespace_slug}/${slugs.memex_slug}/${docTypeToPathSegment(docType)}/${handle}`;
-}
-
-function buildDecisionPath(
-  slugs: MemexSlugs,
-  parentDocType: string,
-  parentHandle: string,
-  decSeq: number,
-): string {
-  return `${buildDocPath(slugs, parentDocType, parentHandle)}/decisions/dec-${decSeq}`;
-}
-
-// Issues hang off a Spec under `/issues/issue-N`, mirroring how decisions hang off
-// `/decisions/dec-N` (spec-112 t-4). The `issue-N` handle is the per-Spec issue seq
-// minted by services/issues.ts — same shape rule as dec-N (renamed from `i-N` per
-// spec-158 dec-3).
-function buildIssuePath(
-  slugs: MemexSlugs,
-  parentDocType: string,
-  parentHandle: string,
-  issueSeq: number,
-): string {
-  return `${buildDocPath(slugs, parentDocType, parentHandle)}/issues/issue-${issueSeq}`;
-}
-
-function kindForDocType(docType: string): MemexSearchKind {
-  if (docType === "spec") return "spec";
-  if (docType === "standard") return "standard";
-  return "document";
-}
-
-function inScopeDocTypes(kind: MemexSearchKind | undefined): string[] | null {
-  if (!kind) return ALL_SECTION_DOC_TYPES;
-  // decision + issue are non-section kinds — their own arms run instead, so the
-  // section query short-circuits to nothing.
-  if (kind === "decision" || kind === "issue") return null;
-  return DOC_TYPES_BY_KIND[kind];
-}
-
-// spec-285: normalise a timestamptz coming back from the driver (Date or ISO
-// string depending on the column path) to a stable ISO-8601 string for the hit
-// shape. null/invalid → null so the formatter and REST JSON degrade gracefully.
-function toIso(value: string | Date | null | undefined): string | null {
-  if (value == null) return null;
-  const d = value instanceof Date ? value : new Date(value);
-  return Number.isNaN(d.getTime()) ? null : d.toISOString();
-}
-
-// spec-259 ac-12: batch-load the open (unresolved) comment count + oldest open
-// comment timestamp for a set of doc ids, in ONE grouped query (no N+1). Keyed
-// by doc_id — a comment transitively belongs to a doc via its section/decision/
-// task target, but doc_id is denormalised on the row (schema comment), so we
-// group on it directly. `resolved_at IS NULL` is the open predicate. Returns a
-// Map doc_id → { count, oldestCreatedAt(ISO) }; absent keys have zero open
-// comments. No comment CONTENT is selected — indicator only.
-interface OpenCommentRow {
-  doc_id: string;
-  open_count: number | string;
-  oldest_created_at: string | Date;
-}
-
-async function loadOpenCommentSummaries(
-  docIds: string[],
-): Promise<Map<string, { count: number; oldestCreatedAt: string }>> {
-  const summaries = new Map<string, { count: number; oldestCreatedAt: string }>();
-  if (docIds.length === 0) return summaries;
-  // De-dupe the id set (decision/section hits can share a parent doc) so the
-  // IN-list is minimal.
-  const uniqueIds = Array.from(new Set(docIds));
-  const idList = sql.join(
-    uniqueIds.map((id) => sql`${id}::uuid`),
-    sql`, `,
-  );
-  const rows = (await db.execute(sql`
-    SELECT
-      doc_id                AS doc_id,
-      COUNT(*)              AS open_count,
-      MIN(created_at)       AS oldest_created_at
-    FROM doc_comments
-    WHERE doc_id IN (${idList})
-      AND resolved_at IS NULL
-    GROUP BY doc_id
-  `)) as unknown as OpenCommentRow[];
-  for (const r of rows) {
-    const count = Number(r.open_count);
-    const oldestCreatedAt = toIso(r.oldest_created_at);
-    if (count > 0 && oldestCreatedAt) {
-      summaries.set(r.doc_id, { count, oldestCreatedAt });
-    }
-  }
-  return summaries;
-}
-
-// spec-259 ac-12: attach the open-comment indicator to each hit by its parent
-// doc. Mutates in place and returns the same array. A decision/section hit
-// inherits its parent doc's open-comment summary (the indicator is doc-scoped).
-async function attachOpenComments(hits: MemexSearchHit[]): Promise<MemexSearchHit[]> {
-  if (hits.length === 0) return hits;
-  const summaries = await loadOpenCommentSummaries(hits.map((h) => h.parentDocId));
-  for (const hit of hits) {
-    const summary = summaries.get(hit.parentDocId);
-    if (summary) hit.openComments = summary;
-  }
-  return hits;
-}
 
 // ── Public entry point ─────────────────────────────────
 
@@ -557,120 +242,6 @@ export async function searchMemex(
   // spec-259 ac-12: enrich the (already-capped) result set with open-comment
   // indicators in one grouped query — batched over the result docs, never N+1.
   return attachOpenComments(merged);
-}
-
-// ── Direct lookup ──────────────────────────────────────
-
-async function lookupByHandle(
-  memexId: string,
-  slugs: MemexSlugs,
-  query: string,
-  includeArchived: boolean,
-): Promise<MemexSearchHit | null> {
-  const archivedClause = includeArchived ? sql`` : sql`AND d.archived_at IS NULL`;
-  const rows = (await db.execute(sql`
-    SELECT
-      s.id            AS section_id,
-      s.section_type  AS section_type,
-      s.title         AS section_title,
-      s.content       AS section_content,
-      s.actor_name    AS section_actor_name,
-      s.updated_at    AS section_updated_at,
-      s.doc_id        AS doc_id,
-      d.handle        AS doc_handle,
-      d.title         AS doc_title,
-      d.status        AS doc_status,
-      d.doc_type      AS doc_type,
-      d.created_at        AS doc_created_at,
-      d.status_changed_at AS doc_status_changed_at,
-      du.name         AS doc_author_name,
-      du.email        AS doc_author_email
-    FROM documents d
-    LEFT JOIN doc_sections s ON s.doc_id = d.id
-    LEFT JOIN users du ON du.id = d.created_by_user_id
-    WHERE d.memex_id = ${memexId}
-      ${archivedClause}
-      AND d.is_demo IS NOT TRUE
-      AND d.handle = ${query.toLowerCase()}
-    ORDER BY s.seq
-  `)) as unknown as HandleRow[];
-
-  if (rows.length === 0) return null;
-
-  const first = rows[0];
-
-  // WHO/WHEN for a handle hit (spec-285): the heading is the DOCUMENT, so
-  // attribute it from the most-recently-updated section (denormalised
-  // actor_name, std-32), falling back to the doc's resolved creator. WHEN is the
-  // latest section's updated_at, else the doc's own status_changed_at/created_at.
-  const sections = rows.filter((r) => r.section_id != null);
-  let latest: HandleRow | null = null;
-  for (const r of sections) {
-    if (!latest || toMillis(r.section_updated_at) > toMillis(latest.section_updated_at)) {
-      latest = r;
-    }
-  }
-  const docAuthorFallback =
-    first.doc_author_name?.trim() || first.doc_author_email?.trim() || null;
-  const sectionActor = latest?.section_actor_name?.trim() || null;
-  const authorName = sectionActor ?? docAuthorFallback;
-  const lastUpdatedAt =
-    toIso(latest?.section_updated_at) ??
-    toIso(first.doc_status_changed_at) ??
-    toIso(first.doc_created_at);
-
-  return {
-    id: first.doc_id,
-    parentDocId: first.doc_id,
-    kind: kindForDocType(first.doc_type),
-    path: buildDocPath(slugs, first.doc_type, first.doc_handle),
-    title: first.doc_title,
-    status: first.doc_status,
-    score: 1,
-    strategies: ["handle"],
-    authorName,
-    lastUpdatedAt,
-    matchingSections: sections.map((r) => ({
-      id: r.section_id,
-      sectionType: r.section_type,
-      title: r.section_title,
-      content: r.section_content,
-      matchedVia: "handle",
-      // spec-259 ac-9: per-section WHO/WHEN. The handle row has no resolved
-      // creator fallback per-section, so use the section's denormalised
-      // actor_name (std-32), else the doc-level resolved creator fallback.
-      authorName: r.section_actor_name?.trim() || docAuthorFallback,
-      lastUpdatedAt: toIso(r.section_updated_at),
-    })),
-  };
-}
-
-// Row shape for lookupByHandle's documents⨝sections⨝users projection (spec-285).
-interface HandleRow {
-  section_id: string;
-  section_type: string;
-  section_title: string | null;
-  section_content: string;
-  section_actor_name: string | null;
-  section_updated_at: string | Date | null;
-  doc_id: string;
-  doc_handle: string;
-  doc_title: string;
-  doc_status: string;
-  doc_type: string;
-  doc_created_at: string | Date | null;
-  doc_status_changed_at: string | Date | null;
-  doc_author_name: string | null;
-  doc_author_email: string | null;
-}
-
-// spec-285: epoch millis for comparing two timestamptz values; missing/invalid
-// sorts oldest so a real timestamp always wins the "latest section" pick.
-function toMillis(value: string | Date | null | undefined): number {
-  if (value == null) return -Infinity;
-  const d = value instanceof Date ? value : new Date(value);
-  const t = d.getTime();
-  return Number.isNaN(t) ? -Infinity : t;
 }
 
 // ── Jump-to lane (spec-64 t-2) ─────────────────────────
@@ -883,690 +454,4 @@ export async function resolveAssignedSpecs(
     }
   }
   return hits;
-}
-
-// ── Section FTS ────────────────────────────────────────
-
-async function runSectionFts(
-  memexId: string,
-  query: string,
-  docTypes: string[],
-  includeArchived: boolean,
-  excludeDocId?: string,
-): Promise<SectionRow[]> {
-  const archivedClause = includeArchived
-    ? sql``
-    : sql`AND d.archived_at IS NULL AND d.paused_at IS NULL`;
-  const excludeClause = excludeDocId
-    ? sql`AND d.id <> ${excludeDocId}::uuid`
-    : sql``;
-  const rows = (await db.execute(sql`
-    SELECT
-      s.id            AS section_id,
-      s.section_type  AS section_type,
-      s.title         AS section_title,
-      s.content       AS section_content,
-      s.doc_id        AS doc_id,
-      s.updated_at    AS updated_at,
-      COALESCE(NULLIF(TRIM(s.actor_name), ''), NULLIF(TRIM(du.name), ''), du.email) AS author_name,
-      d.handle        AS doc_handle,
-      d.title         AS doc_title,
-      d.status        AS doc_status,
-      d.doc_type      AS doc_type,
-      ts_rank(s.content_tsv, plainto_tsquery('english', ${query})) AS rank
-    FROM doc_sections s
-    INNER JOIN documents d ON d.id = s.doc_id
-    LEFT JOIN users du ON du.id = d.created_by_user_id
-    WHERE d.memex_id = ${memexId}
-      AND d.doc_type IN ${sql.raw(`(${docTypes.map((t) => `'${t}'`).join(",")})`)}
-      ${archivedClause}
-      AND d.is_demo IS NOT TRUE
-      ${excludeClause}
-      AND (s.status <> 'deleted' OR s.status IS NULL)
-      AND s.content_tsv @@ plainto_tsquery('english', ${query})
-    ORDER BY rank DESC
-    LIMIT 50
-  `)) as unknown as SectionRow[];
-  return rows;
-}
-
-// ── Section Vector ─────────────────────────────────────
-
-async function runSectionVector(
-  memexId: string,
-  query: string,
-  docTypes: string[],
-  includeArchived: boolean,
-  provider: EmbeddingProvider,
-  maxDistance: number,
-  excludeDocId?: string,
-): Promise<SectionRow[]> {
-  let queryVec: number[];
-  try {
-    [queryVec] = await provider.embed([query], "query");
-  } catch {
-    return [];
-  }
-  if (!queryVec) return [];
-
-  const literal = `[${queryVec.join(",")}]`;
-  const archivedClause = includeArchived
-    ? sql``
-    : sql`AND d.archived_at IS NULL AND d.paused_at IS NULL`;
-  const excludeClause = excludeDocId
-    ? sql`AND d.id <> ${excludeDocId}::uuid`
-    : sql``;
-
-  const rows = (await db.execute(sql`
-    SELECT
-      s.id            AS section_id,
-      s.section_type  AS section_type,
-      s.title         AS section_title,
-      s.content       AS section_content,
-      s.doc_id        AS doc_id,
-      s.updated_at    AS updated_at,
-      COALESCE(NULLIF(TRIM(s.actor_name), ''), NULLIF(TRIM(du.name), ''), du.email) AS author_name,
-      d.handle        AS doc_handle,
-      d.title         AS doc_title,
-      d.status        AS doc_status,
-      d.doc_type      AS doc_type,
-      (s.embedding <=> ${literal}::vector) AS distance
-    FROM doc_sections s
-    INNER JOIN documents d ON d.id = s.doc_id
-    LEFT JOIN users du ON du.id = d.created_by_user_id
-    WHERE d.memex_id = ${memexId}
-      AND d.doc_type IN ${sql.raw(`(${docTypes.map((t) => `'${t}'`).join(",")})`)}
-      ${archivedClause}
-      AND d.is_demo IS NOT TRUE
-      ${excludeClause}
-      AND (s.status <> 'deleted' OR s.status IS NULL)
-      AND s.embedding IS NOT NULL
-      AND s.embedding_model = ${provider.name}
-      AND (s.embedding <=> ${literal}::vector) < ${maxDistance}
-    ORDER BY s.embedding <=> ${literal}::vector
-    LIMIT 50
-  `)) as unknown as SectionRow[];
-  return rows;
-}
-
-// ── Decision FTS ───────────────────────────────────────
-// Inline tsvector since `decisions` doesn't have a generated content_tsv
-// column. Concatenate title + context + resolution at query time. Cost is
-// modest because the table is small relative to doc_sections.
-
-async function runDecisionFts(
-  memexId: string,
-  query: string,
-  includeArchived: boolean,
-  excludeDocId?: string,
-): Promise<DecisionRow[]> {
-  const archivedClause = includeArchived
-    ? sql``
-    : sql`AND d.archived_at IS NULL AND d.paused_at IS NULL`;
-  const excludeClause = excludeDocId
-    ? sql`AND dec.doc_id <> ${excludeDocId}::uuid`
-    : sql``;
-  const rows = (await db.execute(sql`
-    SELECT
-      dec.id          AS decision_id,
-      dec.doc_id      AS doc_id,
-      dec.seq         AS dec_seq,
-      dec.title       AS dec_title,
-      dec.context     AS dec_context,
-      dec.resolution  AS dec_resolution,
-      dec.status      AS dec_status,
-      dec.created_at  AS created_at,
-      COALESCE(NULLIF(TRIM(dec.actor_name), ''), NULLIF(TRIM(au.name), ''), au.email) AS author_name,
-      d.handle        AS doc_handle,
-      d.title         AS doc_title,
-      d.doc_type      AS doc_type,
-      ts_rank(
-        to_tsvector('english',
-          coalesce(dec.title, '') || ' ' ||
-          coalesce(dec.context, '') || ' ' ||
-          coalesce(dec.resolution, '')),
-        plainto_tsquery('english', ${query})
-      ) AS rank
-    FROM decisions dec
-    INNER JOIN documents d ON d.id = dec.doc_id
-    LEFT JOIN users au ON au.id = dec.actor_user_id
-    WHERE dec.memex_id = ${memexId}
-      ${archivedClause}
-      AND d.is_demo IS NOT TRUE
-      ${excludeClause}
-      AND to_tsvector('english',
-            coalesce(dec.title, '') || ' ' ||
-            coalesce(dec.context, '') || ' ' ||
-            coalesce(dec.resolution, ''))
-          @@ plainto_tsquery('english', ${query})
-    ORDER BY rank DESC
-    LIMIT 50
-  `)) as unknown as DecisionRow[];
-  return rows;
-}
-
-// ── Decision Vector ────────────────────────────────────
-
-async function runDecisionVector(
-  memexId: string,
-  query: string,
-  includeArchived: boolean,
-  provider: EmbeddingProvider,
-  maxDistance: number,
-  excludeDocId?: string,
-): Promise<DecisionRow[]> {
-  let queryVec: number[];
-  try {
-    [queryVec] = await provider.embed([query], "query");
-  } catch {
-    return [];
-  }
-  if (!queryVec) return [];
-
-  const literal = `[${queryVec.join(",")}]`;
-  const archivedClause = includeArchived
-    ? sql``
-    : sql`AND d.archived_at IS NULL AND d.paused_at IS NULL`;
-  const excludeClause = excludeDocId
-    ? sql`AND dec.doc_id <> ${excludeDocId}::uuid`
-    : sql``;
-
-  const rows = (await db.execute(sql`
-    SELECT
-      dec.id          AS decision_id,
-      dec.doc_id      AS doc_id,
-      dec.seq         AS dec_seq,
-      dec.title       AS dec_title,
-      dec.context     AS dec_context,
-      dec.resolution  AS dec_resolution,
-      dec.status      AS dec_status,
-      dec.created_at  AS created_at,
-      COALESCE(NULLIF(TRIM(dec.actor_name), ''), NULLIF(TRIM(au.name), ''), au.email) AS author_name,
-      d.handle        AS doc_handle,
-      d.title         AS doc_title,
-      d.doc_type      AS doc_type,
-      (dec.embedding <=> ${literal}::vector) AS distance
-    FROM decisions dec
-    INNER JOIN documents d ON d.id = dec.doc_id
-    LEFT JOIN users au ON au.id = dec.actor_user_id
-    WHERE dec.memex_id = ${memexId}
-      ${archivedClause}
-      AND d.is_demo IS NOT TRUE
-      ${excludeClause}
-      AND dec.embedding IS NOT NULL
-      AND dec.embedding_model = ${provider.name}
-      AND (dec.embedding <=> ${literal}::vector) < ${maxDistance}
-    ORDER BY dec.embedding <=> ${literal}::vector
-    LIMIT 50
-  `)) as unknown as DecisionRow[];
-  return rows;
-}
-
-// ── Issue FTS ──────────────────────────────────────────
-// Same shape as the decision arm (spec-112 t-4). Issues live in their own
-// `issues` table (0068_issues.sql) — bug/todo backlog raised against a Spec —
-// with no generated tsvector column, so we concatenate title + body at query
-// time. Cheap: the issues table is small relative to doc_sections. The join to
-// `documents` is on `iss.doc_id` (the parent Spec), giving us the handle +
-// docType needed to build the `/issues/issue-N` path.
-
-async function runIssueFts(
-  memexId: string,
-  query: string,
-  includeArchived: boolean,
-  excludeDocId?: string,
-): Promise<IssueRow[]> {
-  const archivedClause = includeArchived
-    ? sql``
-    : sql`AND d.archived_at IS NULL AND d.paused_at IS NULL`;
-  const excludeClause = excludeDocId
-    ? sql`AND iss.doc_id <> ${excludeDocId}::uuid`
-    : sql``;
-  const rows = (await db.execute(sql`
-    SELECT
-      iss.id          AS issue_id,
-      iss.doc_id      AS doc_id,
-      iss.seq         AS issue_seq,
-      iss.title       AS issue_title,
-      iss.body        AS issue_body,
-      iss.type        AS issue_type,
-      iss.status      AS issue_status,
-      iss.updated_at  AS updated_at,
-      COALESCE(NULLIF(TRIM(au.name), ''), au.email) AS author_name,
-      d.handle        AS doc_handle,
-      d.title         AS doc_title,
-      d.doc_type      AS doc_type,
-      ts_rank(
-        to_tsvector('english',
-          coalesce(iss.title, '') || ' ' ||
-          coalesce(iss.body, '')),
-        plainto_tsquery('english', ${query})
-      ) AS rank
-    FROM issues iss
-    INNER JOIN documents d ON d.id = iss.doc_id
-    LEFT JOIN users au ON au.id = iss.created_by_user_id
-    WHERE iss.memex_id = ${memexId}
-      ${archivedClause}
-      AND d.is_demo IS NOT TRUE
-      ${excludeClause}
-      AND to_tsvector('english',
-            coalesce(iss.title, '') || ' ' ||
-            coalesce(iss.body, ''))
-          @@ plainto_tsquery('english', ${query})
-    ORDER BY rank DESC
-    LIMIT 50
-  `)) as unknown as IssueRow[];
-  return rows;
-}
-
-// ── Issue Vector ───────────────────────────────────────
-
-async function runIssueVector(
-  memexId: string,
-  query: string,
-  includeArchived: boolean,
-  provider: EmbeddingProvider,
-  maxDistance: number,
-  excludeDocId?: string,
-): Promise<IssueRow[]> {
-  let queryVec: number[];
-  try {
-    [queryVec] = await provider.embed([query], "query");
-  } catch {
-    return [];
-  }
-  if (!queryVec) return [];
-
-  const literal = `[${queryVec.join(",")}]`;
-  const archivedClause = includeArchived
-    ? sql``
-    : sql`AND d.archived_at IS NULL AND d.paused_at IS NULL`;
-  const excludeClause = excludeDocId
-    ? sql`AND iss.doc_id <> ${excludeDocId}::uuid`
-    : sql``;
-
-  const rows = (await db.execute(sql`
-    SELECT
-      iss.id          AS issue_id,
-      iss.doc_id      AS doc_id,
-      iss.seq         AS issue_seq,
-      iss.title       AS issue_title,
-      iss.body        AS issue_body,
-      iss.type        AS issue_type,
-      iss.status      AS issue_status,
-      iss.updated_at  AS updated_at,
-      COALESCE(NULLIF(TRIM(au.name), ''), au.email) AS author_name,
-      d.handle        AS doc_handle,
-      d.title         AS doc_title,
-      d.doc_type      AS doc_type,
-      (iss.embedding <=> ${literal}::vector) AS distance
-    FROM issues iss
-    INNER JOIN documents d ON d.id = iss.doc_id
-    LEFT JOIN users au ON au.id = iss.created_by_user_id
-    WHERE iss.memex_id = ${memexId}
-      ${archivedClause}
-      AND d.is_demo IS NOT TRUE
-      ${excludeClause}
-      AND iss.embedding IS NOT NULL
-      AND iss.embedding_model = ${provider.name}
-      AND (iss.embedding <=> ${literal}::vector) < ${maxDistance}
-    ORDER BY iss.embedding <=> ${literal}::vector
-    LIMIT 50
-  `)) as unknown as IssueRow[];
-  return rows;
-}
-
-// ── RRF merge ──────────────────────────────────────────
-
-interface AccumulatorEntry {
-  id: string;
-  parentDocId: string;
-  kind: MemexSearchKind;
-  path: string;
-  title: string;
-  status: string;
-  score: number;
-  strategies: Set<SearchStrategy>;
-  sectionsByVia: Map<string, MatchingSection>;
-  decisionSnippet?: string;
-  decisionMatchedVia?: SearchStrategy;
-  issueSnippet?: string;
-  issueMatchedVia?: SearchStrategy;
-  issueType?: string;
-  // spec-285 WHO/WHEN. For a doc hit these track the most-recently-updated
-  // matched section (so `authorAt` is the latest-section comparison key); for
-  // decision/issue hits they're set once from the atomic row.
-  authorName: string | null;
-  lastUpdatedAt: string | null;
-  /** Epoch millis of `lastUpdatedAt`, kept only to pick the latest matched
-   *  section across the FTS + vector arms. Not emitted. */
-  authorAtMillis: number;
-}
-
-function pickDecisionSnippet(r: DecisionRow): string {
-  // Prefer resolution → context → title for the snippet body. Cap at 300 chars
-  // (b-34 D-4).
-  const candidate =
-    (r.dec_resolution && r.dec_resolution.trim()) ||
-    (r.dec_context && r.dec_context.trim()) ||
-    r.dec_title;
-  return candidate.length > 300 ? `${candidate.slice(0, 297)}…` : candidate;
-}
-
-function pickIssueSnippet(r: IssueRow): string {
-  // Prefer body → title for the snippet body (an Issue's body carries the
-  // detail; the title is the one-liner). Cap at 300 chars (b-34 D-4).
-  const candidate =
-    (r.issue_body && r.issue_body.trim()) || r.issue_title;
-  return candidate.length > 300 ? `${candidate.slice(0, 297)}…` : candidate;
-}
-
-function mergeWithRrf(
-  sectionFts: SectionRow[],
-  sectionVector: SectionRow[],
-  decisionFts: DecisionRow[],
-  decisionVector: DecisionRow[],
-  issueFts: IssueRow[],
-  issueVector: IssueRow[],
-  slugs: MemexSlugs,
-  limit: number,
-): MemexSearchHit[] {
-  // Sections: keyed by doc_id (group multiple matching sections under one
-  // parent doc). Decisions: keyed by decision_id (atomic).
-  const acc = new Map<string, AccumulatorEntry>();
-
-  function addSectionRows(rows: SectionRow[], via: SearchStrategy): void {
-    for (let i = 0; i < rows.length; i++) {
-      const r = rows[i];
-      const rrfContribution = 1 / (RRF_K + i + 1);
-      let entry = acc.get(`doc:${r.doc_id}`);
-      if (!entry) {
-        entry = {
-          id: r.doc_id,
-          parentDocId: r.doc_id,
-          kind: kindForDocType(r.doc_type),
-          path: buildDocPath(slugs, r.doc_type, r.doc_handle),
-          title: r.doc_title,
-          status: r.doc_status,
-          score: 0,
-          strategies: new Set<SearchStrategy>(),
-          sectionsByVia: new Map(),
-          authorName: null,
-          lastUpdatedAt: null,
-          authorAtMillis: -Infinity,
-        };
-        acc.set(`doc:${r.doc_id}`, entry);
-      }
-      entry.score += rrfContribution;
-      entry.strategies.add(via);
-
-      // WHO/WHEN (spec-285): a doc hit groups many matched sections (and arrives
-      // across the FTS + vector arms in rank order, not time order). Attribute
-      // the doc to its MOST-RECENTLY-UPDATED matched section, so authorName +
-      // lastUpdatedAt answer "who last changed this and when".
-      const rowMillis = toMillis(r.updated_at);
-      if (rowMillis > entry.authorAtMillis) {
-        entry.authorAtMillis = rowMillis;
-        entry.authorName = r.author_name?.trim() || null;
-        entry.lastUpdatedAt = toIso(r.updated_at);
-      }
-
-      // Keep the FIRST `via` that surfaced each section so a section seen by
-      // both FTS and vector reports the higher-confidence search method.
-      if (!entry.sectionsByVia.has(r.section_id)) {
-        entry.sectionsByVia.set(r.section_id, {
-          id: r.section_id,
-          sectionType: r.section_type,
-          title: r.section_title,
-          content: r.section_content,
-          matchedVia: via,
-          // spec-259 ac-9: per-section WHO/WHEN so each matched section's own
-          // creator surfaces, not just the doc-level most-recent one.
-          authorName: r.author_name?.trim() || null,
-          lastUpdatedAt: toIso(r.updated_at),
-        });
-      }
-    }
-  }
-
-  function addDecisionRows(rows: DecisionRow[], via: SearchStrategy): void {
-    for (let i = 0; i < rows.length; i++) {
-      const r = rows[i];
-      const rrfContribution = 1 / (RRF_K + i + 1);
-      let entry = acc.get(`dec:${r.decision_id}`);
-      if (!entry) {
-        entry = {
-          id: r.decision_id,
-          parentDocId: r.doc_id,
-          kind: "decision",
-          path: buildDecisionPath(slugs, r.doc_type, r.doc_handle, r.dec_seq),
-          title: r.dec_title,
-          status: r.dec_status,
-          score: 0,
-          strategies: new Set<SearchStrategy>(),
-          sectionsByVia: new Map(),
-          decisionSnippet: pickDecisionSnippet(r),
-          decisionMatchedVia: via,
-          // WHO/WHEN (spec-285): a decision is atomic, so set once. authorName =
-          // denormalised actor_name (or resolved actor), lastUpdatedAt =
-          // created_at (decisions carry no updated_at — dec-2 fallback).
-          authorName: r.author_name?.trim() || null,
-          lastUpdatedAt: toIso(r.created_at),
-          authorAtMillis: toMillis(r.created_at),
-        };
-        acc.set(`dec:${r.decision_id}`, entry);
-      } else {
-        // Already present from the other arm — preserve original snippet/via
-        // (first-wins), but boost the score.
-      }
-      entry.score += rrfContribution;
-      entry.strategies.add(via);
-    }
-  }
-
-  function addIssueRows(rows: IssueRow[], via: SearchStrategy): void {
-    for (let i = 0; i < rows.length; i++) {
-      const r = rows[i];
-      const rrfContribution = 1 / (RRF_K + i + 1);
-      let entry = acc.get(`iss:${r.issue_id}`);
-      if (!entry) {
-        entry = {
-          id: r.issue_id,
-          parentDocId: r.doc_id,
-          kind: "issue",
-          path: buildIssuePath(slugs, r.doc_type, r.doc_handle, r.issue_seq),
-          title: r.issue_title,
-          status: r.issue_status,
-          score: 0,
-          strategies: new Set<SearchStrategy>(),
-          sectionsByVia: new Map(),
-          issueSnippet: pickIssueSnippet(r),
-          issueMatchedVia: via,
-          issueType: r.issue_type,
-          // WHO/WHEN (spec-285): atomic, set once. Issues have no actor_name, so
-          // authorName is the resolved created_by user; lastUpdatedAt = updated_at.
-          authorName: r.author_name?.trim() || null,
-          lastUpdatedAt: toIso(r.updated_at),
-          authorAtMillis: toMillis(r.updated_at),
-        };
-        acc.set(`iss:${r.issue_id}`, entry);
-      } else {
-        // Already present from the other arm — preserve original snippet/via
-        // (first-wins), but boost the score. Mirrors addDecisionRows.
-      }
-      entry.score += rrfContribution;
-      entry.strategies.add(via);
-    }
-  }
-
-  addSectionRows(sectionFts, "fts");
-  addSectionRows(sectionVector, "vector");
-  addDecisionRows(decisionFts, "fts");
-  addDecisionRows(decisionVector, "vector");
-  addIssueRows(issueFts, "fts");
-  addIssueRows(issueVector, "vector");
-
-  const results: MemexSearchHit[] = Array.from(acc.values()).map((e) => ({
-    id: e.id,
-    kind: e.kind,
-    path: e.path,
-    parentDocId: e.parentDocId,
-    title: e.title,
-    status: e.status,
-    score: e.score,
-    strategies: Array.from(e.strategies).sort(),
-    matchingSections: Array.from(e.sectionsByVia.values()),
-    decisionSnippet: e.decisionSnippet,
-    decisionMatchedVia: e.decisionMatchedVia,
-    issueSnippet: e.issueSnippet,
-    issueMatchedVia: e.issueMatchedVia,
-    issueType: e.issueType,
-    authorName: e.authorName,
-    lastUpdatedAt: e.lastUpdatedAt,
-  }));
-
-  results.sort((a, b) => b.score - a.score);
-  return results.slice(0, limit);
-}
-
-// ══════════════════════════════════════════════════════════
-// Markdown formatter (b-34 D-4)
-// ══════════════════════════════════════════════════════════
-//
-// Renders MemexSearchHit[] to the path-as-heading markdown spec:
-//
-//   ### <canonical-path> — "<title>" (<kind>, <status>)
-//   - Section "<section-title>" (<fts|vector>):
-//     > snippet ≤ 300 chars …
-//
-// For decisions:
-//
-//   ### …/specs/spec-N/decisions/dec-M — "<title>" (decision, <status>)
-//   - (<fts|vector>): <snippet>
-//
-// No UUIDs anywhere (per b-36 D-2/D-7). Score / URL omitted in terse mode;
-// `verbose: true` adds score for debug.
-
-const SNIPPET_MAX_CHARS = 300;
-
-function truncate(text: string, max: number): string {
-  if (text.length <= max) return text;
-  return `${text.slice(0, max - 1)}…`;
-}
-
-export interface FormatOptions {
-  verbose?: boolean;
-  /** When set, hits whose doc id matches (decision hits: parent doc id) get a
-   *  trailing `[current doc]` tag so the agent recognises results that belong
-   *  to the Spec it's currently editing rather than treating them as
-   *  external prior work. Only relevant when search_memex was called with
-   *  `includeCurrentDoc: true` (the default excludes the current doc from
-   *  results entirely, so this never fires). */
-  currentDocId?: string;
-  /** Injectable "now" for relative-age rendering (spec-259 dec-5). Tests pass a
-   *  fixed clock so `timeAgo()` output is deterministic; production omits it and
-   *  the helpers default to the real wall clock. */
-  now?: Date;
-}
-
-// spec-285 + spec-259: the WHO/WHEN byline appended to a hit's heading, e.g.
-// ` · Ryan Soosayraj, 3d ago`. Legible to a human and parseable by an agent
-// (` · <author>, <relative-age>`). Degrades gracefully: author-only, age-only,
-// or nothing at all when neither resolved. spec-259 dec-5 moved the RENDERED
-// timestamp from an absolute YYYY-MM-DD to a relative `timeAgo()` ("Nd ago");
-// the structured `hit.lastUpdatedAt`/`hit.authorName` stay absolute ISO. The
-// author is run through capitalizeDisplayName (spec-259 ac-9) because the SQL
-// WHO-resolver returns raw names that bypass the shared who-resolver. `now` is
-// injectable so unit tests are deterministic.
-function formatWhoWhenByline(
-  authorName: string | null,
-  lastUpdatedAt: string | null,
-  now?: Date,
-): string {
-  const raw = authorName?.trim() || null;
-  const author = raw ? capitalizeDisplayName(raw) : null;
-  const age = lastUpdatedAt ? timeAgo(lastUpdatedAt, now) : null;
-  if (author && age) return ` · ${author}, ${age}`;
-  if (author) return ` · ${author}`;
-  if (age) return ` · ${age}`;
-  return "";
-}
-
-function formatHitByline(hit: MemexSearchHit, now?: Date): string {
-  return formatWhoWhenByline(hit.authorName, hit.lastUpdatedAt, now);
-}
-
-// spec-259 ac-12: the open-comment indicator appended as its own line under a
-// hit, e.g. `- (2 open comments, oldest 3d ago)`. Indicator only — no comment
-// content. Returns null when the hit has zero open comments so the caller emits
-// no line. `now` injectable for deterministic tests.
-function formatOpenCommentsLine(hit: MemexSearchHit, now?: Date): string | null {
-  const oc = hit.openComments;
-  if (!oc || oc.count <= 0) return null;
-  const noun = oc.count === 1 ? "open comment" : "open comments";
-  return `- (${oc.count} ${noun}, oldest ${timeAgo(oc.oldestCreatedAt, now)})`;
-}
-
-function isHitOnCurrentDoc(hit: MemexSearchHit, currentDocId: string): boolean {
-  // parentDocId is the doc UUID for section/doc hits and the parent Spec's
-  // UUID for decision hits. Matching here means the hit belongs to the
-  // Spec the agent is currently editing.
-  return hit.parentDocId === currentDocId;
-}
-
-export function formatSearchResults(
-  query: string,
-  hits: MemexSearchHit[],
-  options: FormatOptions = {},
-): string {
-  if (hits.length === 0) {
-    return `No results for "${query}".`;
-  }
-
-  const verbose = options.verbose === true;
-  const currentDocId = options.currentDocId;
-  const now = options.now;
-  const lines: string[] = [`## Search results for "${query}" (${hits.length} hit${hits.length === 1 ? "" : "s"})`];
-
-  for (const hit of hits) {
-    const scoreSuffix = verbose ? ` (score ${hit.score.toFixed(3)})` : "";
-    const selfTag =
-      currentDocId && isHitOnCurrentDoc(hit, currentDocId) ? " [current doc]" : "";
-    lines.push("");
-    // For issues, fold the bug/todo type into the kind segment of the heading
-    // so a reader can tell a bug from a todo at a glance (spec-112 t-4).
-    const kindLabel =
-      hit.kind === "issue" && hit.issueType ? `issue/${hit.issueType}` : hit.kind;
-    // spec-285: WHO/WHEN byline sits after the (kind, status) segment and before
-    // the [current doc] / verbose-score suffixes, so an agent reading the result
-    // can attribute the hit ("who, when") without opening it.
-    const byline = formatHitByline(hit, now);
-    lines.push(`### ${hit.path} — "${hit.title}" (${kindLabel}, ${hit.status})${byline}${selfTag}${scoreSuffix}`);
-    if (hit.kind === "decision") {
-      const via = hit.decisionMatchedVia ?? "fts";
-      const snippet = hit.decisionSnippet ?? "";
-      lines.push(`- (${via}): ${truncate(snippet, SNIPPET_MAX_CHARS)}`);
-    } else if (hit.kind === "issue") {
-      const via = hit.issueMatchedVia ?? "fts";
-      const snippet = hit.issueSnippet ?? "";
-      lines.push(`- (${via}): ${truncate(snippet, SNIPPET_MAX_CHARS)}`);
-    } else {
-      for (const sec of hit.matchingSections) {
-        const titleSeg = sec.title ? `"${sec.title}"` : `(${sec.sectionType})`;
-        const snippet = truncate((sec.content ?? "").trim(), SNIPPET_MAX_CHARS);
-        // spec-259 ac-9: per-section WHO/WHEN so a multi-section doc hit surfaces
-        // each matched section's own creator + relative age, not just the
-        // doc-level latest. Same capitalize + timeAgo treatment as the heading.
-        const secByline = formatWhoWhenByline(sec.authorName, sec.lastUpdatedAt, now);
-        lines.push(`- Section ${titleSeg} (${sec.matchedVia})${secByline}:`);
-        lines.push(`  > ${snippet}`);
-      }
-    }
-    // spec-259 ac-12: one lightweight open-comment indicator line per hit (after
-    // the snippet block). Rendered only when the hit has open comments.
-    const openCommentsLine = formatOpenCommentsLine(hit, now);
-    if (openCommentsLine) lines.push(openCommentsLine);
-  }
-
-  return lines.join("\n");
 }
